fix(sidecar): return version in bolt_metadata endpoint
